CRITICAL UPDATE:
THIS PROJECT HAS BEEN TEMPORARILY SHELVED DUE TO LACK OF FINANCES. I don't know when I will be able to secure enough funding to continue building the drone, but it is unlike to be anytime in the next few months.

Once I have the necessary funds, I will resume work on this project and provide updates accordingly. This latest update contains an AI demonstration of what a simplified alternative of YMFC-AL-flight controller could be.
NOTE: YMFC is a project credits to Joop Brokking, I am not the original creator of YMFC, I am merely adapting it for my own use and learning purposes. Find the original code and tutorial on his website: https://www.brokking.net/